This merge combines:
- tinyxml2 migration from claude/upgrade-to-tinyxml2 branch
- Improved Makefile structure and code improvements from master

Key changes:
- Updated Makefile to use tinyxml2 instead of tinerxml wrapper
- Removed tinerxml.cpp from build sources
- Kept tinyxml2 includes in level.cpp and client.cpp
- Integrated master's improvements to entity, renderer, and other files

+++ src/ui.h
@@ -25,7 +25,7 @@ public:
 // ButtonArray class (should probably be called ButtonList)
 class ButtonArray {
 public:
-    unsigned short Buttons[15]; // 16 max buttons per array
+    int Buttons[15]; // 16 max buttons per array
     int ButtonCount;
     std::string Name;
     bool Draw;
